% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_de_deelstaten.R
\docType{data}
\name{sp_de_deelstaten}
\alias{sp_de_deelstaten}
\title{Map outlining the German 'deelstaten'}
\format{An object of class SpatialPolygonsDataFrame with 16 rows and 4 columns.}
\source{
Shapefile data online available \href{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-2-500-000-stand-01-01-vg2500.html}{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-2-500-000-stand-01-01-vg2500.html}
}
\usage{
sp_de_deelstaten
}
\description{
Germany has 16 'deelstaten' or Bundesländer. Read more on these administrative areas in Germany on \href{https://nl.wikipedia.org/wiki/Deelstaten_van_Duitsland}{Wikipedia}.
}
\keyword{datasets}
