% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_de_districten.R
\docType{data}
\name{sp_de_districten}
\alias{sp_de_districten}
\title{Map outlining the German districts}
\format{An object of class SpatialPolygonsDataFrame with 401 rows and 4 columns.}
\source{
Shapefile data online available \href{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-2-500-000-stand-01-01-vg2500.html}{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-2-500-000-stand-01-01-vg2500.html}
}
\usage{
sp_de_districten
}
\description{
Germany has 401 districts or 'Kreise'. Read more on these administrative areas in Germany on \href{https://nl.wikipedia.org/wiki/Landkreis}{Wikipedia}.
}
\keyword{datasets}
