% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_de_regeringsdistricten.R
\docType{data}
\name{sp_de_regeringsdistricten}
\alias{sp_de_regeringsdistricten}
\title{Map outlining the German 'regeringsdistricten'}
\format{An object of class SpatialPolygonsDataFrame with 401 rows and 4 columns.}
\source{
Shapefile data online available \href{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-2-500-000-stand-01-01-vg2500.html}{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-2-500-000-stand-01-01-vg2500.html}
}
\usage{
sp_de_regeringsdistricten
}
\description{
Germany has 19 'regeringsdistricten' or 'Regierungsbezirk'. These are areas in Germany that contain multiple districts or Kreise. Read more on these administrative areas in Germany on \href{https://nl.wikipedia.org/wiki/Regierungsbezirk}{Wikipedia}.
}
\keyword{datasets}
